# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Senac Bike Store API - By Condé",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        },
        {
            "url": "INSIRA AQUI A URL do Host",
            "description": "Produção"
        }
    ],
    "paths": {
...
        "/api/clientes/{id}/pedidos": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Retorna a Lista de Pedidos.",
                "description": "Retorna o JSON da Lista de Pedidos.",
                "operationId": "getPedidosList",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Cliente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ordenacao",
                        "in": "query",
                        "description": "Ordernação dos dados",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidosResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidosResponseException500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/clientes/{1}/pedidos": {
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Armazena os dados de um novo Pedido.",
                "description": "Armazena e retorna o JSON de um novo Pedido.",
                "operationId": "postPedido",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidoNewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException500"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pedidos/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Retorna os dados de um Pedido.",
                "description": "Retorna o JSON de um Pedido.",
                "operationId": "getPedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Apaga os dados de um pedido.",
                "description": "Apaga e retorna o JSON de confirmação.",
                "operationId": "deletePedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Atualiza os dados de um pedido.",
                "description": "Atualiza e retorna o JSON de um Pedido.",
                "operationId": "patchPedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do Pedido",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidoNewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação executada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ocorreu um erro interno.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException500"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pedido não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidoResponseException404"
                                }
                            }
                        }
                    },
                    "406": {
                        "description": "O campo Id deve ser numérico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaResponseException406"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PedidoNewRequest": {
                "title": "PedidoNewRequest",
                "description": "Requisição enviada via Body para uma inserção ou atualização de um novo Pedido",
                "required": [
                    "nome_da_Marca"
                ],
                "properties": {
                    "data_requisitada": {
                        "title": "data_requisitada",
                        "description": "Data do pedido.",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
            "PedidoResponse": {
                "title": "PedidoResponse",
                "description": "Resposta enviada via JSON para os dados de um Pedido",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Pedido retornado."
                    },
                    "pedido": {
                        "$ref": "#/components/schemas/Pedido"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PedidoResponse"
                }
            },
...
            "PedidosResponse": {
                "title": "PedidosResponse",
                "description": "Resposta enviada via JSON para Lista de Pedidos",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Lista de pedidos retornada."
                    },
                    "pedidos": {
                        "title": "pedidos",
                        "description": "Lista de Pedidos retornada.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Pedido"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PedidosResponse"
                }
            },
            "PedidosResponseException500": {
                "title": "PedidosResponseException500",
                "description": "Resposta enviada via JSON para Lista de Pedidos",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "HTTP Status.",
                        "type": "integer",
                        "format": "int64",
                        "example": 500
                    },
                    "mensagem": {
                        "title": "mensagem",
                        "description": "Mensagem informativa da resposta.",
                        "type": "string",
                        "format": "string",
                        "example": "Erro Interno."
                    },
                    "pedidos": {
                        "title": "produtos",
                        "description": "Lista de Pedidos retornada.",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PedidosResponseException500"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": {
        "bearer": {
            "type": "http",
            "description": "Authorization token obtained from logging in.",
            "name": "Authorization",
            "in": "header",
            "scheme": "bearer"
        }
    }
}